ipynb file
Context -
DRS bank is facing challenging times. Their NPAs (Non-Performing Assets) has been on a rise recently and a large part of these are due to the loans given to individual customers(borrowers). Chief Risk Officer of the bank decides to put in a scientifically robust framework for approval of loans to individual customers to minimize the risk of loans converting into NPAs and initiates a project for the data science team at the bank. You, as a senior member of the team, are assigned this project.

Objective -
To identify the criteria to approve loans for an individual customer such that the likelihood of the loan delinquency is minimized

Key questions to be answered –
What are the factors that drive the behavior of loan delinquency?

Dataset:
ID: Customer ID
isDelinquent : indicates whether the customer is delinquent or not (1 => Yes, 0 => No)
term: Loan term in months
gender: Gender of the borrower
age: Age of the borrower
purpose: Purpose of Loan
home_ownership: Status of borrower's home
FICO: FICO (i.e. the bureau score) of the borrower
Domain Information –
Transactor – A person who pays his due amount balance full and on time. Revolver – A person who pays the minimum due amount but keeps revolving his balance and does not pay the full amount. Delinquent - Delinquency means that you are behind on payments, a person who fails to pay even the minimum due amount. Defaulter – Once you are delinquent for a certain period your lender will declare you to be in the default stage. Risk Analytics – A wide domain in the financial and banking industry, basically analyzing the risk of the customer.